requirements.txt
et_xmlfile==2.0.0
MouseInfo==0.1.3
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
